// packages/ui/src/constants/colors.d.ts
declare const Colors: {
    coral: string;
    yellow: string;
    green: string;
    white: string;
    freshGreen: string;
    naverGreen: string;
    facebookBlue: string;
    amber: string;
    lightishGreen: string;
    lightTeal: string;
    tealish: string;
    turtleGreen: string;
    darkCoral: string;
    rosyPink: string;
    mango: string;
    transparent: string;
    sunYellow: string;
    error: string;
    black3: string;
    black20: string;
    black24: string;
    red: string;
    coralPink: string;
    lightCoralPink: string;
    secondary: string;
    blue6Alpha: string;
    paleGray: string;
    gray2C: string;
    grayD8: string;
    gray4E: string;
    gray9E: string;
    gray77: string;
    gray69: string;
    gray86: string;
    blueGray0: string;
    blueGray1: string;
    blueGray2: string;
    blueGray3: string;
    blueGray4: string;
    blueGray5: string;
    blueGray6: string;
    blueGray7: string;
    blueGray8: string;
    blueGray9: string;
    blue0: string;
    blue1: string;
    blue2: string;
    blue3: string;
    blue4: string;
    blue5: string;
    blue6: string;
    blue7: string;
    blue8: string;
    blue9: string;
    violet0: string;
    violet1: string;
    violet2: string;
    violet3: string;
    violet4: string;
    violet5: string;
    violet6: string;
    violet7: string;
    violet8: string;
    violet9: string;
    normalGray1: string;
    normalGray2: string;
    normalGray3: string;
    normalGray4: string;
    normalGray5: string;
    normalGray6: string;
    normalGray7: string;
    normalGray8: string;
    normalGray9: string;
    gray0: string;
    gray1: string;
    gray2: string;
    gray3: string;
    gray4: string;
    gray5: string;
    gray6: string;
    gray7: string;
    gray8: string;
    gray9: string;
};
export default Colors;